Increase receiver chain count from 5 -> 10
When we ratchet, we can immediately use the new sending chain safely,
however, we have to retain some number of old receiving chain to be able
to decrypt any in-flight messages, or messages that arrive out of order
(i.e.  messages that were *sent* before we ratcheted, but not received
until after we ratcheted).

The tradeoff is storage size (how many old chains we keep around per
session) vs resilience (how many ratchet steps do we maintain the
ability to decrypt with).

Usually message delivery is quick and serial and this isn't an issue.
However sometimes message delays of several minutes have been observed.
If this delay occurs *while* the parties are engaged in an active
back-and-forth, sending messages every few seconds, it's very likely
we'll ratchet past the threshold and see bad decrypted messages.

// FREEBIE